fix: change datatype of pin

+++ include/MyLora.h
@@ -7,14 +7,14 @@
 class MyLora
 {
 private:
-    int _nssPin;
-    int _resetPin;
-    int _dio0Pin;
+    uint8_t _nssPin;
+    uint8_t _resetPin;
+    uint8_t _dio0Pin;
     byte _localAddress;
     String parseData(String dataString);
 
 public:
-    MyLora(const int *nssPin, const int *resetPin, const int *dio0Pin, const byte *localAddress);
+    MyLora(const byte *nssPin, const byte *resetPin, const byte *dio0Pin, const byte *localAddress);
     ~MyLora();
     void initilize(const long frequency);
     void sendMessage(byte destination, String message);
